+++ tests/plugin_test/src/167_trait_crash.php
@@ -1,0 +1,16 @@
+<?php
+
+// Regression test for https://github.com/phan/phan/issues/3796
+namespace TraitCrash;
+/**
+ * @method diff(\DateTimeInterface $date)
+ */
+trait BaseTrait {
+    public function diff($date = null) {
+        return 1;
+    }
+}
+
+trait Other {
+    use BaseTrait;
+}